+++ st2client/st2client/commands/action.py
@@ -19,6 +19,7 @@ import copy
 import json
 import logging
 import textwrap
+import calendar
 import time
 import six
 import sys
@@ -33,6 +34,7 @@ from st2client.formatters import table
 from st2client.formatters import execution as execution_formatter
 from st2client.utils import jsutil
 from st2client.utils.date import format_isodate
+from st2client.utils.date import parse as parse_isotime
 from st2client.utils.color import format_status
 
 LOG = logging.getLogger(__name__)
@@ -117,6 +119,31 @@ def format_wf_instances(instances):
         else:
             instance.id = NON_WF_PREFIX + instance.id
     return instances
+
+
+def format_execution_statuses(instances):
+    result = []
+    for instance in instances:
+        instance = format_execution_status(instance)
+        result.append(instance)
+
+    return result
+
+
+def format_execution_status(instance):
+    """
+    Augment instance "status" attribute with number of seconds which have elapsed for all the
+    executions which are in running state.
+    """
+    if instance.status == LIVEACTION_STATUS_RUNNING and instance.start_timestamp:
+        start_timestamp = instance.start_timestamp
+        start_timestamp = parse_isotime(start_timestamp)
+        start_timestamp = calendar.timegm(start_timestamp.timetuple())
+        now = int(time.time())
+        elapsed_seconds = (now - start_timestamp)
+        instance.status = '%s (%ss elapsed)' % (instance.status, elapsed_seconds)
+
+    return instance
 
 
 class ActionBranch(resource.ResourceBranch):
@@ -924,6 +951,11 @@ class ActionExecutionListCommand(ActionExecutionReadCommand):
 
     def run_and_print(self, args, **kwargs):
         instances = format_wf_instances(self.run(args, **kwargs))
+
+        if not args.json:
+            # Include elapsed time for running executions
+            instances = format_execution_statuses(instances)
+
         self.print_output(reversed(instances), table.MultiColumnTable,
                           attributes=args.attr, widths=args.width,
                           json=args.json,
@@ -962,6 +994,10 @@ class ActionExecutionGetCommand(ActionRunCommandMixin, ActionExecutionReadComman
     def run_and_print(self, args, **kwargs):
         try:
             execution = self.run(args, **kwargs)
+
+            if not args.json:
+                # Include elapsed time for running executions
+                execution = format_execution_status(execution)
         except resource.ResourceNotFoundError:
             self.print_not_found(args.id)
             raise OperationFailureException('Execution %s not found.' % (args.id))